Components: Ensure that SlotFill does not use portals in React Native
