+++ packages/components/src/slot-fill/index.js
@@ -1,19 +1,13 @@
 /**
  * Internal dependencies
  */
-import BaseSlot from './slot';
 import BaseFill from './fill';
-import Provider from './context';
-import BubblesVirtuallySlot from './bubbles-virtually/slot';
+import BaseSlot from './slot';
 import BubblesVirtuallyFill from './bubbles-virtually/fill';
+import BubblesVirtuallySlot from './bubbles-virtually/slot';
+import BubblesVirtuallySlotFillProvider from './bubbles-virtually/slot-fill-provider';
+import SlotFillProvider from './provider';
 import useSlot from './bubbles-virtually/use-slot';
-
-export function Slot( { bubblesVirtually, ...props } ) {
-	if ( bubblesVirtually ) {
-		return <BubblesVirtuallySlot { ...props } />;
-	}
-	return <BaseSlot { ...props } />;
-}
 
 export function Fill( props ) {
 	// We're adding both Fills here so they can register themselves before
@@ -24,6 +18,23 @@ export function Fill( props ) {
 			<BaseFill { ...props } />
 			<BubblesVirtuallyFill { ...props } />
 		</>
+	);
+}
+
+export function Slot( { bubblesVirtually, ...props } ) {
+	if ( bubblesVirtually ) {
+		return <BubblesVirtuallySlot { ...props } />;
+	}
+	return <BaseSlot { ...props } />;
+}
+
+export function Provider( { children, ...props } ) {
+	return (
+		<SlotFillProvider { ...props }>
+			<BubblesVirtuallySlotFillProvider>
+				{ children }
+			</BubblesVirtuallySlotFillProvider>
+		</SlotFillProvider>
 	);
 }
 
@@ -40,4 +51,4 @@ export function createSlotFill( name ) {
 	};
 }
 
-export { useSlot, Provider };
+export { useSlot };
