+++ packages/components/src/slot-fill/slot.js
@@ -16,7 +16,7 @@ import {
 /**
  * Internal dependencies
  */
-import { Consumer } from './context';
+import SlotFillContext from './context';
 
 class SlotComponent extends Component {
 	constructor() {
@@ -86,7 +86,7 @@ class SlotComponent extends Component {
 }
 
 const Slot = ( props ) => (
-	<Consumer>
+	<SlotFillContext.Consumer>
 		{ ( { registerSlot, unregisterSlot, getFills } ) => (
 			<SlotComponent
 				{ ...props }
@@ -95,7 +95,7 @@ const Slot = ( props ) => (
 				getFills={ getFills }
 			/>
 		) }
-	</Consumer>
+	</SlotFillContext.Consumer>
 );
 
 export default Slot;
